First assignment
Senior year
Coded in Python
1-1: Given a text file, it will process the contents and find word
frequencies; at the end, it prints out these frequent words in
descending order along with the count. On an AWS instance, it was able
to process a 1 GB file in 3 minutes.
1-2: Similar to 1-1, but it is given 2 files and it outputs the words
they have in common and the total count of words they have in common